+++ scripts/post-build.js
@@ -1,12 +1,9 @@
 const fs = require('fs');
-const glob = require("glob");
-const terser = require("terser");
+// const glob = require("glob");
+// const terser = require("terser");
 
-const [,,mode] = process.argv;
+// const [,,mode] = process.argv;
 copyAssets();
-if (mode !== 'dev') {
-    minify();
-}
 
 function copyAssets() {
     const {scripts, devDependencies, ['lint-staged']: _, config, husky, ...cleanPackage} = JSON.parse(fs.readFileSync('package.json').toString());
@@ -14,9 +11,9 @@ function copyAssets() {
     fs.copyFileSync('README.md', 'dist/README.md');
 }
 
-function minify() {
-    glob.sync('dist/**/*.js').forEach(async (filePath) => {
-        const {code} = await terser.minify(fs.readFileSync(filePath, "utf8"));
-        fs.writeFileSync(filePath, code);
-    });
-}
+// function minify() {
+//     glob.sync('dist/**/*.js').forEach(async (filePath) => {
+//         const {code} = await terser.minify(fs.readFileSync(filePath, "utf8"));
+//         fs.writeFileSync(filePath, code);
+//     });
+// }