+++ src/keys-detective/get-translation-files-path.ts
@@ -1,10 +1,9 @@
-import * as glob from 'glob';
-
 import { FileFormats } from '../types';
+import { normalizedGlob } from '../utils/normalize-glob-path';
 
 export function getTranslationFilesPath(
   path: string,
   fileFormat: FileFormats
 ): string[] {
-  return glob.sync(`${path}/**/*.${fileFormat}`);
+  return normalizedGlob(`${path}/**/*.${fileFormat}`);
 }